refactor(analytics): split analytics types facade for GH727
