+++ src/core/analytics/types_tests.rs
@@ -1,0 +1,10 @@
+use super::*;
+
+#[path = "types_tests/cost_tests.rs"]
+mod cost_tests;
+#[path = "types_tests/request_tests.rs"]
+mod request_tests;
+#[path = "types_tests/usage_tests.rs"]
+mod usage_tests;
+#[path = "types_tests/workflow_tests.rs"]
+mod workflow_tests;
